+++ CHANGELOG.md
@@ -16,4 +16,6 @@
   to store an ArrayBuffer, that's fine, your usage of this
   won't change, it's only that internally anything you pass
   to it is always packaged as a payload.
+
+- adjusted 'info' endpoint to match channel server